Add id for the bar container
Thank you for this plugin. Really simple to use :)

I think adding an Id to the bar container would allow easy customization with CSS (for example to change the position of the bar => above bottom navbar on mobile, or to hide it easily by adding a hidden class if the mobile menu is open)

What do you think about it?